Tile stores Piece object instead of name
Tile class now stores the piece object instead of piece short name. During selection of pieces (before the start of the game), when a player selects a piece and clicks select button, the piece object is created and stored in the Player class and the name of that piece is removed from the selection list.